add feature to trigger managed rolling decommission
This adds the ability to do planned decommissioning of StatefulSet pods.

To trigger a managed decommission, set an annotation on the Cluster resource with an RFC3339 deadline.

```
kubectl annotate cluster managed-decommission operator.redpanda.com/managed-decommission="2999-12-31T00:00:00Z"
```

The Cluster controller will decommission each pod from 0 to N, delete the pvcs and the pod, waiting for the replacement pod to join and the cluster to be healthy between each deletion.

When finished, the controller will remove the annotation.

Should the deadline pass during the managed decommission, the controller will not continue decommissioning pods. To resume the decommission, set the annotation to a timestamp in the future

+++ src/go/k8s/pkg/utils/kubernetes.go
@@ -29,7 +29,7 @@ func IsPodReady(pod *corev1.Pod) bool {
 	return false
 }
 
-func GetPodOrdinal(podName, clusterName string) (int64, error) {
+func GetPodOrdinal(podName, clusterName string) (int32, error) {
 	// Pod name needs to have at least 2 more characters
 	if len(podName) < len(clusterName)+2 {
 		return -1, fmt.Errorf("pod name (%s) and cluster name (%s): %w", podName, clusterName, ErrInvalidInputParameters)
@@ -41,5 +41,5 @@ func GetPodOrdinal(podName, clusterName string) (int64, error) {
 	if err != nil {
 		return -1, fmt.Errorf("parsing int failed (%s): %w", ordinalStr, err)
 	}
-	return ordinal, nil
+	return int32(ordinal), nil
 }
